chore: 폴더명 컴포넌트 명이랑 동일하게 변경
